Split integration test support by domain
