+++ tests/grpc.rs
@@ -1,7 +1,29 @@
 mod support;
 
 use sha2::{Digest as Sha2Digest, Sha256};
-use support::*;
+use std::io::Write;
+use std::net::Ipv4Addr;
+use std::process::{Command, Stdio};
+use std::sync::mpsc;
+use std::time::Duration;
+use support::common::{
+    FetchOpts, assert_exit, fetch_bin, run_fetch, run_fetch_opts, start_read_capture, temp_file,
+    wait_child,
+};
+use support::dns::start_udp_dns_server;
+#[cfg(unix)]
+use support::grpc::start_delayed_response_grpc_h2c_server;
+use support::grpc::{
+    grpc_frame, proto_field_string, proto_field_varint, start_delayed_bidi_grpc_h2c_server,
+    start_reflection_grpc_h2c_many_small_frames_server, start_reflection_grpc_h2c_server,
+    start_reflection_grpc_h2c_v1_error_response_server, start_reflection_grpc_tls_server,
+    start_reflection_grpc_tls_server_with_versions, start_status_grpc_h2c_server,
+    write_health_descriptor_set, write_stream_descriptor_set,
+};
+use support::http::{TestResponse, TestServer};
+#[cfg(unix)]
+use support::pty::{configure_pty_child, open_pty, start_pty_capture};
+use tempfile::TempDir;
 
 fn sha256_hex(bytes: &[u8]) -> String {
     const HEX: &[u8; 16] = b"0123456789abcdef";
